*change:[KRF-00001] Change main component name and props
